Move all functionality related to the zeroing UI section -> zero.py
